+++ src/card/cards/combat_constraints.rs
@@ -208,3 +208,5 @@ pub const INTO_THE_WILDS: CardDefinitionId = CardDefinitionId(2007);
 pub const TRAUMATIZE: CardDefinitionId = CardDefinitionId(2008);
 pub const DEATHS_PRESENCE: CardDefinitionId = CardDefinitionId(2009);
 pub const AJANI_CALLER_OF_THE_PRIDE: CardDefinitionId = CardDefinitionId(2010);
+pub const SPELL_RUPTURE: CardDefinitionId = CardDefinitionId(2011);
+pub const GIANT_ADEPHAGE: CardDefinitionId = CardDefinitionId(2012);