+++ src/game/tests.rs
@@ -543,6 +543,7 @@ mod blocking_prevention;
 mod blocking_relation;
 mod blocking_relationship;
 mod board_conditioned_statics;
+mod board_sized_counter;
 mod candelabra_of_tawnos;
 mod cannot_attack;
 mod cannot_block;